requirements.txt
pandas
psycopg2-binary
python-dotenv
tabulate
google-generativeai
anthropic
gspread
google-auth
plotly
altair
bcrypt
